fix some haddock problems
Unfortunately it seem (emperically) that haddock only allows the pipe
character on the first line of document, not on every line. Or more
literally, it allows them, but treats them as part of the output to
display in the rendered document.

Fixing that, it also wasn't happy about not introducing the code block
as a new paragraph.